% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBCS.R
\name{runBCS}
\alias{runBCS}
\title{Run the \code{behaviourchange} Shiny app}
\usage{
runBCS()
}
\description{
Run the \code{behaviourchange} Shiny app
}
\examples{
behaviourchange::runBCS()

}
\author{
Ty Stanford and Dot Dumuid
}
